Fix the location-handling
When no location was set an exception was thrown. And when a location
was set the styling of the frontend was difficult because the possible
options were not sufficient. Therefore an flag was added to ignore the
location and the code was modified to prevent the exception for an empty
location.